refactor(MTV-696): address PR review feedback on populator throttling
- Extract ForkliftController populator helpers to forklift_controller_populator.py
- Use ResourceEditor context manager for automatic CR restore
- Move wait_for_plan_secret to copyoffload_plan_secret.py to break circular import
- Move _ACTIVE_POPULATOR_POD_PHASES to module-level constants
- Clarify available_replicas precedence in deployment wait logic
